Moved utils file up a level, making it semantically accessible to all models
